Restructuring the dotfiles
In order to keep home folders clean, this repo keeps the configuration
files for important programs. They can then be symlinked in the
homefolder. This ensures consitent updates, and that all changes are
visible to the repository. When the configuration changes on one host,
it's just an update on all other hosts.

Next up is a script that automates the symlinking