Creating username and machine name from functions
When running in ssh sessions it is of interest to know which user I am
on the remote machine, along with the name of that remote machine. While
that is not yet implemented, it is prepared.

Getting usernames and machine names from a function could also be used
for things like getting the username for the identity I use in a
webservice that the current repo belongs to, e.g. change user@machine to
HaMster21@GitHub in a GitHub repo.